Bug fix and re-add external processing code (#9)
* Lidarr Process change to artist folder name

Previously the process would only import downloads if the Lidarr Artist name, matched the downloaded artist folder name. This alters that to tell Lidarr to scan for downloads based on the actual downloaded folder name. This fixes some artists that would not previously import due to differences in the Lidarr & Folder name

* fix

fixing error

* Enable Lidarr Import Processing for "artist" download method

This will enable importing only wanted albums using the artist download method. Lidarr will not currently import albums/artists that are not already in Lidarr

* Move enable lidarr process setting to main for support of artist mode

* Move s Enable Lidarr Processing setting because it works for both "artist" and "wanted" modes

* Update config.sample

* Options to support new functions

Adding a feature to remove unwanted files

* Updated for new features

* Adding additional cleanup features and code cleanup

* Add new settings for External Download Processing

* re-arrange settings

* Add new settings to readme

* Bug Fix, external processing code re-added after merge